+++ src/dockery/main.py
@@ -1,4 +1,5 @@
 import docker
+from docker import errors
 from docker.models.containers import Container
 from docker.models.volumes import Volume
 from docker.models.images import Image
@@ -7,6 +8,7 @@ import click
 from rich import print
 
 from .gui import AppGUI
+from .utils import var_dump
 
 default_options = [
     click.option(
@@ -16,6 +18,7 @@ default_options = [
     ),
     click.option("--ssh", is_flag=True),
     click.option("--api-version", default="1.35", help="Docker API version"),
+    click.option("--format", "-f", default="yaml", help="Output format"),
 ]
 
 
@@ -56,7 +59,7 @@ def main(ctx, **kargs):
 @add_options(default_options)
 def df(**kargs):
     client = get_client(**kargs)
-    print(client.df())
+    var_dump(client.df(), kargs["format"])
 
 
 @click.command
@@ -64,8 +67,8 @@ def df(**kargs):
 def volumes(**kargs):
     client = get_client(**kargs)
     vlms: list[Volume] = client.volumes.list()  # type: ignore
-    for v in vlms:
-        print(v.attrs)
+    vlm_list = list(map(lambda x: x.attrs, vlms))
+    var_dump(vlm_list, kargs["format"])
 
 
 @click.command
@@ -74,10 +77,10 @@ def volumes(**kargs):
 def ps(**kargs):
     client = get_client(**kargs)
     containers: list[Container] = client.containers.list(
-        all=kargs["all"],
+        all=kargs["all"]
     )  # type: ignore
-    for c in containers:
-        print(c.attrs)
+    con_list = list(map(lambda x: x.attrs, containers))
+    var_dump(con_list, kargs["format"])
 
 
 @click.command
@@ -85,10 +88,8 @@ def ps(**kargs):
 def stats(**kargs):
     client = get_client(**kargs)
     containers: list[Container] = client.containers.list(all=True)  # type: ignore
-    for c in containers:
-        print(f"[b]{c.name}:")
-        stats = c.stats(decode=None, stream=False)
-        print(stats)
+    stats = list(map(lambda x: x.stats(stream=False), containers))
+    var_dump(stats, kargs["format"])
 
 
 @click.command
@@ -96,16 +97,37 @@ def stats(**kargs):
 def images(**kargs):
     client = get_client(**kargs)
     imgs: list[Image] = client.images.list()  # type: ignore
-    for i in imgs:
-        print(i.attrs)
+    img_list = list(map(lambda x: x.attrs, imgs))
+    var_dump(img_list, kargs["format"])
+
 
 @click.command
 @add_options(default_options)
 def networks(**kargs):
     client = get_client(**kargs)
     netw: list[Network] = client.networks.list()  # type: ignore
-    for i in netw:
-        print(i.attrs)
+    net_list = list(map(lambda x: x.attrs, netw))
+    var_dump(net_list, kargs["format"])
+
+
+@click.command
+@add_options(default_options)
+@click.argument("container")
+@click.option("--stream", is_flag=True)
+def logs(**kargs):
+    client = get_client(**kargs)
+    try:
+        container: Container = client.containers.get(kargs["container"])  # type: ignore
+    except errors.NotFound:
+        print(f"Container [b]{kargs['container']}[/] not found")
+    else:
+        if kargs["stream"]:
+            for log in container.logs(stream=True):
+                print(log.decode("utf-8", errors="ignore"), end="")
+        else:
+            logs: bytes = container.logs()
+            print(logs.decode("utf-8", errors="ignore"))
+
 
 main.add_command(df)
 main.add_command(volumes)
@@ -113,6 +135,7 @@ main.add_command(ps)
 main.add_command(stats)
 main.add_command(images)
 main.add_command(networks)
+main.add_command(logs)
 
 if __name__ == "__main__":
     main()