+++ src/dockery/utils.py
@@ -1,3 +1,12 @@
+import json
+from typing import Any, Literal
+from rich.console import Console
+from rich.syntax import Syntax
+import yaml
+
+console = Console()
+
+
 def get_cpu_usage(stats: dict) -> float:
     cpu_stats = stats["cpu_stats"]
     precpu_stats = stats["precpu_stats"]
@@ -28,3 +37,12 @@ def get_mem_usage(stats: dict) -> float:
     limit = stats["memory_stats"]["limit"]
     percentage = mem_used / limit * 100
     return percentage
+
+
+def var_dump(obj: Any, syntax: Literal["json", "yaml"] = "yaml"):
+    if syntax == "yaml":
+        text_obj = yaml.safe_dump(obj, indent=2)
+    elif syntax == "json":
+        text_obj = json.dumps(obj, default=str, indent=2)
+    out = Syntax(text_obj, syntax, theme="ansi_dark")
+    console.print(out)