+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## :zap: New Features
 
-- Nothing new :disappointed:
+- New `logs` command, use it like this:
+
+    `dockery logs {container}`, add the parameter
+    `--stream` to get the logs in real time.
+
+- New `--format` parameter to specify the output format of almost all commands output, for now it supports `json` and `yaml` only.
 
 ## :lady_beetle: Fixes
 
-- Fix error when images has not tags
+- Avoid exceptions on logs decode